backend/requirements.txt
# asgiref==3.5.0
# cffi==1.15.0
# sqlparse==0.4.2

bcrypt==3.2.0
Django==4.0.3
django-ses==3.0.1
django-debug-toolbar==3.2.4
drf-nested-routers==0.93.4
djangorestframework==3.13.1
PyJWT==2.3.0
requests==2.27.1
boto3==1.21.35
django-storages==1.12.3
Pillow==9.0.1
psycopg2-binary==2.9.3
gunicorn==19.10.0
python-dotenv==0.20.0
django-cors-headers==3.5.0
pydash==6.0.0
greatawesomeutils==0.0.27
pandas
beautifulsoup4==4.11.1
joblib